refactor: move generated services to match app layout
